Trait web3::api::eth::Eth

pub trait Eth {
    fn accounts(&self) -> Result<Vec<String>>;
    fn block_number(&self) -> Result<String>;
    fn call(&self,
        options: CallRequest,
        block: Option<BlockNumber>)
        -> Result<Bytes>;
    fn coinbase(&self) -> Result<String>;
    fn compile_lll(&self, String) -> Result<Bytes>;
    fn compile_solidity(&self, String) -> Result<Bytes>;
    fn compile_serpent(&self, String) -> Result<Bytes>;
    fn estimate_gas(&self,
                options: CallRequest,
                block: Option<BlockNumber>)
                -> Result<String>;
    fn gas_price(&self) -> Result<String>;
    fn balance(&self, address: String, block: Option<BlockNumber>) -> Result<String>;
    fn block(&self, block: BlockId, include_txs: bool) -> Result<()>;
    fn block_transaction_count(&self, block: BlockId) -> Result<Option<String>>;
    fn code(&self, address: String, block: Option<BlockNumber>) -> Result<Bytes>;
    fn compilers(&self) -> Result<Vec<String>>;
    fn storage(&self,
           address: String,
           idx: String,
           block: Option<BlockNumber>)
           -> Result<String>;
    fn transaction_count(&self,
                     address: String,
                     block: Option<BlockNumber>)
                     -> Result<String>;
    fn transaction(&self, id: TransactionId) -> Result<Option<()>>;
    fn transaction_receipt(&self, hash: String) -> Result<Option<()>>;
    fn uncle(&self, block: BlockId, index: String) -> Result<Option<()>>;
    fn uncle_count(&self, block: BlockId) -> Result<Option<String>>;
    fn work(&self) -> Result<()>;
    fn hashrate(&self) -> Result<String>;
    fn mining(&self) -> Result<bool>;
    fn new_block_filter(&self) -> Result<String>;
    fn new_pending_transaction_filter(&self) -> Result<String>;
    fn protocol_version(&self) -> Result<String>;
    fn send_raw_transaction(&self, rlp: Bytes) -> Result<String>;
    fn send_transaction(&self, tx: TransactionRequest) -> Result<String>;
    fn sign(&self, address: String, data: Bytes) -> Result<String>;
    fn submit_hashrate(&self, rate: String, id: String) -> Result<bool>;
    fn submit_work(&self,
               nonce: String,
               pow_hash: String,
               mix_hash: String)
               -> Result<bool>;
    fn syncing(&self) -> Result<bool>;
}

List of methods from eth namespace

Required Methods

fn accounts(&self) -> Result<Vec<String>>

Get list of available accounts.

fn block_number(&self) -> Result<String>

Get current block number

fn call(&self,
options: CallRequest,
block: Option<BlockNumber>)
-> Result<Bytes>

Call a constant method of contract without changing the state of the blockchain.

fn coinbase(&self) -> Result<String>

Get coinbase address

fn compile_lll(&self, String) -> Result<Bytes>

Compile LLL

fn compile_solidity(&self, String) -> Result<Bytes>

Compile Solidity

fn compile_serpent(&self, String) -> Result<Bytes>

Compile Serpent

fn estimate_gas(&self,
options: CallRequest,
block: Option<BlockNumber>)
-> Result<String>

Call a contract without changing the state of the blockchain to estimate gas usage.

fn gas_price(&self) -> Result<String>

Get current recommended gas price

fn balance(&self, address: String, block: Option<BlockNumber>) -> Result<String>

Get balance of given address

fn block(&self, block: BlockId, include_txs: bool) -> Result<()>

Get block details

fn block_transaction_count(&self, block: BlockId) -> Result<Option<String>>

Get number of transactions in block

fn code(&self, address: String, block: Option<BlockNumber>) -> Result<Bytes>

Get code under given address

fn compilers(&self) -> Result<Vec<String>>

Get supported compilers

fn storage(&self,
address: String,
idx: String,
block: Option<BlockNumber>)
-> Result<String>

Get storage entry

fn transaction_count(&self,
address: String,
block: Option<BlockNumber>)
-> Result<String>

Get nonce

fn transaction(&self, id: TransactionId) -> Result<Option<()>>

Get transaction

fn transaction_receipt(&self, hash: String) -> Result<Option<()>>

Get transaction receipt

fn uncle(&self, block: BlockId, index: String) -> Result<Option<()>>

Get uncle

fn uncle_count(&self, block: BlockId) -> Result<Option<String>>

Get uncle count in block

fn work(&self) -> Result<()>

Get work package

fn hashrate(&self) -> Result<String>

Get hash rate

fn mining(&self) -> Result<bool>

Get mining status

fn new_block_filter(&self) -> Result<String>

Start new block filter

fn new_pending_transaction_filter(&self) -> Result<String>

Start new pending transaction filter

fn protocol_version(&self) -> Result<String>

Start new pending transaction filter

fn send_raw_transaction(&self, rlp: Bytes) -> Result<String>

Sends a rlp-encoded signed transaction

fn send_transaction(&self, tx: TransactionRequest) -> Result<String>

Sends a transaction transaction

fn sign(&self, address: String, data: Bytes) -> Result<String>

Signs a hash of given data

fn submit_hashrate(&self, rate: String, id: String) -> Result<bool>

Submit hashrate of external miner

fn submit_work(&self,
nonce: String,
pow_hash: String,
mix_hash: String)
-> Result<bool>

Submit work of external miner

fn syncing(&self) -> Result<bool>

Get syncing status