Struct web3::api::ParitySet

pub struct ParitySet<T> { /* fields omitted */ }

Parity_Set Specific API

Methods

impl<T: Transport> ParitySet<T>

pub fn accept_non_reserved_peers(&self) -> CallFuture<bool, T::Out>

Set Parity to accept non-reserved peers (default behavior)

pub fn add_reserved_peer(&self, enode: &str) -> CallFuture<bool, T::Out>

Add a reserved peer

pub fn drop_non_reserved_peers(&self) -> CallFuture<bool, T::Out>

Set Parity to drop all non-reserved peers. To restore default behavior call parity_acceptNonReservedPeers

pub fn execute_upgrade(&self) -> CallFuture<bool, T::Out>

Attempts to upgrade Parity to the version specified in parity_upgradeReady

pub fn hash_content(&self, url: &str) -> CallFuture<H256, T::Out>

Creates a hash of a file at a given URL

pub fn remove_reserved_peer(&self, enode: &str) -> CallFuture<bool, T::Out>

Remove a reserved peer

pub fn set_author(&self, author: &Address) -> CallFuture<bool, T::Out>

Changes author (coinbase) for mined blocks

pub fn set_chain(&self, chain: &str) -> CallFuture<bool, T::Out>

Sets the network spec file Parity is using

pub fn set_engine_signer(
&self,
address: &Address,
password: &str
) -> CallFuture<bool, T::Out>

Sets an authority account for signing consensus messages

pub fn set_extra_data(&self, data: &H256) -> CallFuture<bool, T::Out>

Changes extra data for newly mined blocks

pub fn set_gas_ceil_target(&self, quantity: &H256) -> CallFuture<bool, T::Out>

Sets new gas ceiling target for mined blocks

pub fn set_gas_floor_target(&self, quantity: &H256) -> CallFuture<bool, T::Out>

Sets a new gas floor target for mined blocks

pub fn set_max_transaction_gas(
&self,
quantity: &H256
) -> CallFuture<bool, T::Out>

Sets the maximum amount of gas a single transaction may consume

pub fn set_min_gas_price(&self, quantity: &H256) -> CallFuture<bool, T::Out>

Changes minimal gas price for transaction to be accepted to the queue

pub fn set_mode(&self, mode: &str) -> CallFuture<bool, T::Out>

Changes the operating mode of Parity.

pub fn set_transactions_limit(&self, limit: &H256) -> CallFuture<bool, T::Out>

Changes limit for transactions in queue. (NOT WORKING !)

pub fn upgrade_ready(&self) -> CallFuture<Option<String>, T::Out>

Returns a ReleaseInfo object describing the release which is available for upgrade or null if none is available.