Struct web3::api::Personal

pub struct Personal<T> { /* fields omitted */ }

Personal namespace

Methods

impl<T: Transport> Personal<T>

fn list_accounts(&self) -> CallResult<Vec<Address>, T::Out>

Returns a list of available accounts.

fn new_account(&self, password: &str) -> CallResult<Address, T::Out>

Creates a new account and protects it with given password. Returns the address of created account.

fn unlock_account(
&self,
address: Address,
password: &str,
duration: Option<u16>
) -> CallResult<bool, T::Out>

Unlocks the account with given password for some period of time (or single transaction). Returns true if the call was successful.

fn send_transaction(
&self,
transaction: TransactionRequest,
password: &str
) -> CallResult<H256, T::Out>

Sends a transaction from locked account. Returns transaction hash.