Struct web3::api::Web3

pub struct Web3<T: Transport> { /* fields omitted */ }

Web3 wrapper for all namespaces

Methods

impl<T: Transport> Web3<T>

pub fn new(transport: T) -> Self

Create new Web3 with given transport

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn transport(&self) -> &T

Borrows a transport.

pub fn api<A: Namespace<T>>(&self) -> A

Access methods from custom namespace

pub fn eth(&self) -> Eth<T>

Access methods from eth namespace

pub fn net(&self) -> Net<T>

Access methods from net namespace

pub fn web3(&self) -> Web3<T>

Access methods from web3 namespace

pub fn eth_filter(&self) -> EthFilter<T>

Access filter methods from eth namespace

pub fn personal(&self) -> Personal<T>

Access methods from personal namespace

pub fn wait_for_confirmations<F, V>(
&self,
poll_interval: Duration,
confirmations: usize,
check: V
) -> Confirmations<T, V, F::Future> where
F: IntoFuture<Item = Option<U256>, Error = Error>,
V: ConfirmationCheck<Check = F>,

Should be used to wait for confirmations

pub fn send_transaction_with_confirmation(
&self,
tx: TransactionRequest,
poll_interval: Duration,
confirmations: usize
) -> SendTransactionWithConfirmation<T>

Sends transaction and returns future resolved after transaction is confirmed

pub fn send_raw_transaction_with_confirmation(
&self,
tx: Bytes,
poll_interval: Duration,
confirmations: usize
) -> SendTransactionWithConfirmation<T>

Sends raw transaction and returns future resolved after transaction is confirmed

impl<T: DuplexTransport> Web3<T>

pub fn eth_subscribe(&self) -> EthSubscribe<T>

Access subscribe methods from eth namespace