Module web3::confirm

Easy to use utilities for confirmations.

Structs

Confirmations
On each new block checks confirmations.

SendTransactionWithConfirmation
Sends transaction and then checks if has been confirmed.

Traits

ConfirmationCheck
Checks whether an event has been confirmed.

Functions

send_raw_transaction_with_confirmation
Sends raw transaction and returns future resolved after transaction is confirmed

send_transaction_with_confirmation
Sends transaction and returns future resolved after transaction is confirmed

wait_for_confirmations
Should be used to wait for confirmations