Trait web3::confirm::ConfirmationCheck

pub trait ConfirmationCheck {
    type Check: Future<Output = Result<Option<U64>>>;
    fn check(&self) -> Self::Check;
}

Checks whether an event has been confirmed.

Associated Types

type Check: Future<Output = Result<Option<U64>>>

Future resolved when is known whether an event has been confirmed.

Required methods

fn check(&self) -> Self::Check

Should be called to get future which resolves when confirmation state is known.