Struct web3::contract::ens::Ens

pub struct Ens<T: Transport> { /* private fields */ }

Ethereum Name Service interface.

Implementations

impl<T: Transport> Ens<T>

pub async fn owner(&self, node: &str) -> Result<Address, Error>

Returns the owner of the name specified by node.

pub async fn resolver(&self, node: &str) -> Result<Address, Error>

Returns the address of the resolver responsible for the name specified by node.

pub async fn ttl(&self, node: &str) -> Result<u64, Error>

Returns the caching time-to-live of the name specified by node.

Systems that wish to cache information about a name, including ownership, resolver address, and records, should respect this value.

If TTL is zero, new data should be fetched on each query.

pub async fn set_owner(
&self,
from: Address,
node: &str,
owner: Address
) -> Result<TransactionId, Error>

Reassigns ownership of the name identified by node to owner.

Only callable by the current owner of the name.

Emits the following event:

event Transfer(bytes32 indexed node, address owner);

pub async fn set_resolver(
&self,
from: Address,
node: &str,
address: Address
) -> Result<TransactionId, Error>

Updates the resolver associated with the name identified by node to resolver.

Only callable by the current owner of the name. resolver must specify the address of a contract that implements the Resolver interface.

Emits the following event:

event NewResolver(bytes32 indexed node, address resolver);

pub async fn set_ttl(
&self,
from: Address,
node: &str,
ttl: u64
) -> Result<TransactionId, Error>

Updates the caching time-to-live of the name identified by node.

Only callable by the current owner of the name.

Emits the following event:

event NewTTL(bytes32 indexed node, uint64 ttl);

pub async fn set_subdomain_owner(
&self,
from: Address,
node: &str,
label: &str,
owner: Address
) -> Result<TransactionId, Error>

Creates a new subdomain of node, assigning ownership of it to the specified owner.

If the domain already exists, ownership is reassigned but the resolver and TTL are left unmodified.

For example, if you own alice.eth and want to create the subdomain iam.alice.eth, supply alice.eth as the node, and iam as the label.

Emits the following event:

event NewOwner(bytes32 indexed node, bytes32 indexed label, address owner);

pub async fn set_record(
&self,
from: Address,
node: &str,
owner: Address,
resolver: Address,
ttl: u64
) -> Result<TransactionId, Error>

Sets the owner, resolver, and TTL for an ENS record in a single operation.

This function is offered for convenience, and is exactly equivalent to calling set_resolver, set_ttl and set_owner in that order.

pub async fn set_subdomain_record(
&self,
from: Address,
node: &str,
label: &str,
owner: Address,
resolver: Address,
ttl: u64
) -> Result<TransactionId, Error>

Sets the owner, resolver and TTL for a subdomain, creating it if necessary.

This function is offered for convenience, and permits setting all three fields without first transferring ownership of the subdomain to the caller.

pub async fn set_approval_for_all(
&self,
from: Address,
operator: Address,
approved: bool
) -> Result<TransactionId, Error>

Sets or clears an approval.

Approved accounts can execute all ENS registry operations on behalf of the caller.

pub async fn is_approved_for_all(
&self,
owner: Address,
operator: Address
) -> Result<bool, Error>

Returns true if operator is approved to make ENS registry operations on behalf of owner.

pub async fn record_exists(&self, node: &str) -> Result<bool, Error>

Returns true if node exists in this ENS registry.

This will return false for records that are in the legacy ENS registry but have not yet been migrated to the new one.

pub async fn supports_interface(
&self,
node: &str,
interface_id: [u8; 4]
) -> Result<bool, Error>

ENS uses ERC 165 for interface detection.

ERC 165 requires that supporting contracts implement a function, supportsInterface, which takes an interface ID and returns a boolean value indicating if this interface is supported or not. Interface IDs are calculated as the exclusive-or of the four-byte function identifiers of each function included in the interface.

For example, addr(bytes32) has the function ID 0x3b3b57de. Because it is the only function in the Ethereum Address interface, its interface ID is also 0x3b3b57de, and so calling supportsInterface(0x3b3b57de) will return true for any resolver that supports addr(). ERC 165 has an interface ID of 0x01ffc9a7, so supportsInterface(0x01ffc9a7) will always return true for any ERC 165 supporting contract (and hence for any resolver).

Note that the public resolver does not expose explicit interfaces for setter functions, so there are no automated means to check for support for a given setter function.

pub async fn eth_address(&self, node: &str) -> Result<Address, Error>

Returns the Ethereum address associated with the provided node, or 0 if none.

This function has interface ID 0x3b3b57de.

This function is specified in EIP 137.

pub async fn set_eth_address(
&self,
from: Address,
node: &str,
addr: Address
) -> Result<TransactionId, Error>

Sets the Ethereum address associated with the provided node to addr.

Only callable by the owner of node.

Emits the following event:

event AddrChanged(bytes32 indexed node, address a);

pub async fn blockchain_address(
&self,
node: &str,
coin_type: U256
) -> Result<Vec<u8>, Error>

Returns the Blockchain address associated with the provided node and coin_type, or 0 if none.

This function has interface ID 0xf1cb7e06.

This function is specified in EIP 2304.

The return value is the cryptocurency address in its native binary format and each blockchain address has a different encoding and decoding method.

For example, the Bitcoin address 1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa base58check decodes to the 21 bytes 0062e907b15cbf27d5425399ebf6f0fb50ebb88f18 then scriptPubkey encodes to 25 bytes 76a91462e907b15cbf27d5425399ebf6f0fb50ebb88f1888ac whereas the BNB address bnb1grpf0955h0ykzq3ar5nmum7y6gdfl6lxfn46h2 Bech32 decodes to the binary representation 40c2979694bbc961023d1d27be6fc4d21a9febe6.

A zero-length string will be returned if the specified coin ID does not exist on the specified node.

pub async fn set_blockchain_address(
&self,
from: Address,
node: &str,
coin_type: U256,
addr: Vec<u8>
) -> Result<TransactionId, Error>

Sets the blockchain address associated with the provided node and coin_type to addr.

coinType is the cryptocurrency coin type index from SLIP44.

Only callable by the owner of node.

Emits the following event:

event AddressChanged(bytes32 indexed node, uint coinType, bytes newAddress);

pub async fn pubkey(&self, node: &str) -> Result<([u8; 32], [u8; 32]), Error>

Returns the ECDSA SECP256k1 public key for node, as a 2-tuple (x, y). If no public key is set, (0, 0) is returned.

This function has interface ID 0xc8690233.

This function is specified in EIP 619.

pub async fn set_pubkey(
&self,
from: Address,
node: &str,
x: [u8; 32],
y: [u8; 32]
) -> Result<TransactionId, Error>

Sets the ECDSA SECP256k1 public key for node to (x, y).

Only callable by the owner of node.

Emits the following event:

event PubkeyChanged(bytes32 indexed node, bytes32 x, bytes32 y);

pub async fn content_hash(&self, node: &str) -> Result<Vec<u8>, Error>

Returns the content hash for node, if one exists.

Values are formatted as machine-readable multicodecs, as specified in EIP 1577.

content_hash is used to store IPFS and Swarm content hashes, which permit resolving ENS addresses to distributed content (eg, websites) hosted on these distributed networks.

This function has interface ID 0xbc1c58d1.

This function is specified in EIP 1577.

pub async fn set_content_hash(
&self,
from: Address,
node: &str,
hash: Vec<u8>
) -> Result<TransactionId, Error>

Sets the content hash for the provided node to hash.

Only callable by the owner of node.

Values are formatted as machine-readable multicodecs, as specified in EIP 1577.

Emits the following event:

event ContenthashChanged(bytes32 indexed node, bytes hash);

pub async fn text(&self, node: &str, key: String) -> Result<String, Error>

Retrieves text metadata for node. Each name may have multiple pieces of metadata, identified by a unique string key. If no text data exists for node with the key key, the empty string is returned.

Standard values for key are:

key	Meaning

email	An email address

url	A URL

avatar	A URL to an image used as an avatar or logo

description	A description of the name

notice	A notice regarding this name

keywords	A list of comma-separated keywords, ordered by most significant first; clients that interpresent this field may choose a threshold beyond which to ignore

In addition, anyone may specify vendor-specific keys, which must be prefixed with vnd.. The following vendor-specific keys are currently known:

key	Meaning

com.twitter	Twitter handle

com.github	Github username

This function has interface ID 0x59d1d43c.

This function is specified in EIP 634.

pub async fn set_text(
&self,
from: Address,
node: &str,
key: String,
value: String
) -> Result<TransactionId, Error>

Sets text metadata for node with the unique key key to value, overwriting anything previously stored for node and key. To clear a text field, set it to the empty string.

Only callable by the owner of node.

Emits the following event:

event TextChanged(bytes32 indexed node, string indexedKey, string key);

pub async fn canonical_name(&self, addr: Address) -> Result<String, Error>

Returns the canonical ENS name associated with the provided addr. Used exclusively for reverse resolution.

This function has interface ID 0x691f3431.

This function is specified in EIP 181.

pub async fn set_canonical_name(
&self,
from: Address,
node: &str,
name: String
) -> Result<TransactionId, Error>

Sets the canonical ENS name for the provided node to name.

Only callable by the owner of node.

Emits the following event:

event NameChanged(bytes32 indexed node, string name);