Module web3::helpers

Web3 helpers.

Structs

CallFuture
Calls decode on the result of the wrapped future.

Functions

build_request
Build a JSON-RPC request.

decode
Takes any type which is deserializable from rpc::Value and such a value and yields the deserialized value

serialize
Serialize a type. Panics if the type is returns error during serialization.

to_notification_from_slice
Parse bytes slice into JSON-RPC notification.

to_response_from_slice
Parse bytes slice into JSON-RPC response.

to_result_from_output
Parse rpc::Output into Result.

to_results_from_outputs
Parse a Vec of rpc::Output into Result.

to_string
Serializes a request to string. Panics if the type returns error during serialization.