Module web3::signing

Signing capabilities and utilities.

Structs

SecretKeyRef

A SecretKey reference wrapper.

Signature

A struct that represents the components of a secp256k1 signature.

Enums

RecoveryError

Error during sender recovery.

SigningError

Error during signing.

Traits

Key

A trait representing ethereum-compatible key with signing capabilities.

Functions

hash_message

Hash a message according to EIP-191.

keccak256

Compute the Keccak-256 hash of input bytes.

namehash

Compute the hash of a domain name using the namehash algorithm.

recover

Recover a sender, given message and the signature.

Type Definitions

NameHash

Result of the name hash algotithm.