Struct web3::transports::http::Http

pub struct Http { /* fields omitted */ }

HTTP Transport (synchronous)

Methods

impl Http

pub fn new(url: &str) -> Result<(EventLoopHandle, Self)>

Create new HTTP transport with given URL and spawn an event loop in a separate thread. NOTE: Dropping event loop handle will stop the transport layer!

pub fn with_max_parallel(
url: &str,
max_parallel: usize
) -> Result<(EventLoopHandle, Self)>

Create new HTTP transport with given URL and spawn an event loop in a separate thread. You can set a maximal number of parallel requests using the second parameter. NOTE: Dropping event loop handle will stop the transport layer!

pub fn with_event_loop(
url: &str,
handle: &Handle,
max_parallel: usize
) -> Result<Self>

Create new HTTP transport with given URL and existing event loop handle.