Module web3::api

Web3 implementation

Structs

BaseFilter
Base filter handle. Uninstall filter on drop. Allows to poll the filter.

CreateFilter
Future which resolves with new filter

Eth
Eth namespace

EthFilter
Eth namespace, filters

FilterStream
Stream of events

Net
Net namespace

Personal
Personal namespace

SubscriptionId
ID of subscription returned from eth_subscribe

SubscriptionStream
Stream of notifications from a subscription Given a type deserializable from rpc::Value and a subscription id, yields items of that type as notifications are delivered.

Web3
Web3 wrapper for all namespaces

Web3Api
Web3 namespace

Traits

Namespace
Common API for all namespaces