Module web3::api

Web3 implementation

Structs

Accounts
Accounts namespace

BaseFilter
Base filter handle. Uninstall filter on drop. Allows to poll the filter.

Eth
Eth namespace

EthFilter
Eth namespace, filters

EthSubscribe
Eth namespace, subscriptions

Net
Net namespace

Parity
Parity namespace

ParityAccounts
Parity_Accounts namespace

ParitySet
Parity_Set Specific API

Personal
Personal namespace

SubscriptionId
ID of subscription returned from eth_subscribe

SubscriptionStream
Stream of notifications from a subscription Given a type deserializable from rpc::Value and a subscription id, yields items of that type as notifications are delivered.

Traces
Trace namespace

Txpool
Txpool namespace

Web3
Web3 wrapper for all namespaces

Web3Api
Web3 namespace

Traits

Namespace
Common API for all namespaces