Struct web3::api::Eth

pub struct Eth<T> { /* fields omitted */ }

Eth namespace

Methods

impl<T: Transport> Eth<T>

pub fn accounts(&self) -> CallFuture<Vec<Address>, T::Out>

Get list of available accounts.

pub fn block_number(&self) -> CallFuture<U64, T::Out>

Get current block number

pub fn call(
&self,
req: CallRequest,
block: Option<BlockNumber>
) -> CallFuture<Bytes, T::Out>

Call a constant method of contract without changing the state of the blockchain.

pub fn coinbase(&self) -> CallFuture<Address, T::Out>

Get coinbase address

pub fn compile_lll(&self, code: String) -> CallFuture<Bytes, T::Out>

Compile LLL

pub fn compile_solidity(&self, code: String) -> CallFuture<Bytes, T::Out>

Compile Solidity

pub fn compile_serpent(&self, code: String) -> CallFuture<Bytes, T::Out>

Compile Serpent

pub fn estimate_gas(
&self,
req: CallRequest,
block: Option<BlockNumber>
) -> CallFuture<U256, T::Out>

Call a contract without changing the state of the blockchain to estimate gas usage.

pub fn gas_price(&self) -> CallFuture<U256, T::Out>

Get current recommended gas price

pub fn balance(
&self,
address: Address,
block: Option<BlockNumber>
) -> CallFuture<U256, T::Out>

Get balance of given address

pub fn logs(&self, filter: Filter) -> CallFuture<Vec<Log>, T::Out>

Get all logs matching a given filter object

pub fn block(&self, block: BlockId) -> CallFuture<Option<Block<H256>>, T::Out>

Get block details with transaction hashes.

pub fn block_with_txs(
&self,
block: BlockId
) -> CallFuture<Option<Block<Transaction>>, T::Out>

Get block details with full transaction objects.

pub fn block_transaction_count(
&self,
block: BlockId
) -> CallFuture<Option<U256>, T::Out>

Get number of transactions in block

pub fn code(
&self,
address: Address,
block: Option<BlockNumber>
) -> CallFuture<Bytes, T::Out>

Get code under given address

pub fn compilers(&self) -> CallFuture<Vec<String>, T::Out>

Get supported compilers

pub fn chain_id(&self) -> CallFuture<U256, T::Out>

Get chain id

pub fn storage(
&self,
address: Address,
idx: U256,
block: Option<BlockNumber>
) -> CallFuture<H256, T::Out>

Get storage entry

pub fn transaction_count(
&self,
address: Address,
block: Option<BlockNumber>
) -> CallFuture<U256, T::Out>

Get nonce

pub fn transaction(
&self,
id: TransactionId
) -> CallFuture<Option<Transaction>, T::Out>

Get transaction

pub fn transaction_receipt(
&self,
hash: H256
) -> CallFuture<Option<TransactionReceipt>, T::Out>

Get transaction receipt

pub fn uncle(
&self,
block: BlockId,
index: Index
) -> CallFuture<Option<Block<H256>>, T::Out>

Get uncle by block ID and uncle index -- transactions only has hashes.

pub fn uncle_count(&self, block: BlockId) -> CallFuture<Option<U256>, T::Out>

Get uncle count in block

pub fn work(&self) -> CallFuture<Work, T::Out>

Get work package

pub fn hashrate(&self) -> CallFuture<U256, T::Out>

Get hash rate

pub fn mining(&self) -> CallFuture<bool, T::Out>

Get mining status

pub fn new_block_filter(&self) -> CallFuture<U256, T::Out>

Start new block filter

pub fn new_pending_transaction_filter(&self) -> CallFuture<U256, T::Out>

Start new pending transaction filter

pub fn protocol_version(&self) -> CallFuture<String, T::Out>

Start new pending transaction filter

pub fn send_raw_transaction(&self, rlp: Bytes) -> CallFuture<H256, T::Out>

Sends a rlp-encoded signed transaction

pub fn send_transaction(
&self,
tx: TransactionRequest
) -> CallFuture<H256, T::Out>

Sends a transaction transaction

pub fn sign(&self, address: Address, data: Bytes) -> CallFuture<H520, T::Out>

Signs a hash of given data

pub fn submit_hashrate(&self, rate: U256, id: H256) -> CallFuture<bool, T::Out>

Submit hashrate of external miner

pub fn submit_work(
&self,
nonce: H64,
pow_hash: H256,
mix_hash: H256
) -> CallFuture<bool, T::Out>

Submit work of external miner

pub fn syncing(&self) -> CallFuture<SyncState, T::Out>

Get syncing status