Struct web3::api::Eth

pub struct Eth<T> { /* fields omitted */ }

Eth namespace

Methods

impl<T: Transport> Eth<T>

fn accounts(&self) -> CallResult<Vec<Address>, T::Out>

Get list of available accounts.

fn block_number(&self) -> CallResult<U256, T::Out>

Get current block number

fn call(
&self,
req: CallRequest,
block: Option<BlockNumber>
) -> CallResult<Bytes, T::Out>

Call a constant method of contract without changing the state of the blockchain.

fn coinbase(&self) -> CallResult<Address, T::Out>

Get coinbase address

fn compile_lll(&self, code: String) -> CallResult<Bytes, T::Out>

Compile LLL

fn compile_solidity(&self, code: String) -> CallResult<Bytes, T::Out>

Compile Solidity

fn compile_serpent(&self, code: String) -> CallResult<Bytes, T::Out>

Compile Serpent

fn estimate_gas(
&self,
req: CallRequest,
block: Option<BlockNumber>
) -> CallResult<U256, T::Out>

Call a contract without changing the state of the blockchain to estimate gas usage.

fn gas_price(&self) -> CallResult<U256, T::Out>

Get current recommended gas price

fn balance(
&self,
address: Address,
block: Option<BlockNumber>
) -> CallResult<U256, T::Out>

Get balance of given address

fn block(&self, block: BlockId) -> CallResult<Block<H256>, T::Out>

Get block details with transaction hashes.

fn block_with_txs(
&self,
block: BlockId
) -> CallResult<Block<Transaction>, T::Out>

Get block details with full transaction objects.

fn block_transaction_count(
&self,
block: BlockId
) -> CallResult<Option<U256>, T::Out>

Get number of transactions in block

fn code(
&self,
address: Address,
block: Option<BlockNumber>
) -> CallResult<Bytes, T::Out>

Get code under given address

fn compilers(&self) -> CallResult<Vec<String>, T::Out>

Get supported compilers

fn storage(
&self,
address: Address,
idx: U256,
block: Option<BlockNumber>
) -> CallResult<H256, T::Out>

Get storage entry

fn transaction_count(
&self,
address: Address,
block: Option<BlockNumber>
) -> CallResult<U256, T::Out>

Get nonce

fn transaction(
&self,
id: TransactionId
) -> CallResult<Option<Transaction>, T::Out>

Get transaction

fn transaction_receipt(
&self,
hash: H256
) -> CallResult<Option<TransactionReceipt>, T::Out>

Get transaction receipt

fn uncle(
&self,
block: BlockId,
index: Index
) -> CallResult<Option<Block<H256>>, T::Out>

Get uncle by block ID and uncle index -- transactions only has hashes.

fn uncle_count(&self, block: BlockId) -> CallResult<Option<U256>, T::Out>

Get uncle count in block

fn work(&self) -> CallResult<Work, T::Out>

Get work package

fn hashrate(&self) -> CallResult<U256, T::Out>

Get hash rate

fn mining(&self) -> CallResult<bool, T::Out>

Get mining status

fn new_block_filter(&self) -> CallResult<U256, T::Out>

Start new block filter

fn new_pending_transaction_filter(&self) -> CallResult<U256, T::Out>

Start new pending transaction filter

fn protocol_version(&self) -> CallResult<String, T::Out>

Start new pending transaction filter

fn send_raw_transaction(&self, rlp: Bytes) -> CallResult<H256, T::Out>

Sends a rlp-encoded signed transaction

fn send_transaction(&self, tx: TransactionRequest) -> CallResult<H256, T::Out>

Sends a transaction transaction

fn sign(&self, address: Address, data: Bytes) -> CallResult<H520, T::Out>

Signs a hash of given data

fn submit_hashrate(&self, rate: U256, id: H256) -> CallResult<bool, T::Out>

Submit hashrate of external miner

fn submit_work(
&self,
nonce: H64,
pow_hash: H256,
mix_hash: H256
) -> CallResult<bool, T::Out>

Submit work of external miner

fn syncing(&self) -> CallResult<bool, T::Out>

Get syncing status