Struct web3::contract::Contract

pub struct Contract<T: Transport> { /* fields omitted */ }

Ethereum Contract Interface

Methods

impl<T: Transport> Contract<T>

fn new(eth: Eth<T>, address: Address, abi: Contract) -> Self

Creates new Contract Interface given blockchain address and ABI

fn from_json(eth: Eth<T>, address: Address, json: &[u8]) -> Result<Self, Error>

Creates new Contract Interface given blockchain address and JSON containing ABI

fn call<P>(
&self,
func: &str,
params: P,
from: Address,
options: Options
) -> QueryResult<H256, T::Out> where
P: Tokenize,

Execute a contract function

fn estimate_gas<P>(
&self,
func: &str,
params: P,
from: Address,
options: Options
) -> QueryResult<U256, T::Out> where
P: Tokenize,

Estimate gas required for this function call.

fn query<R, A, P>(
&self,
func: &str,
params: P,
from: A,
options: Options
) -> QueryResult<R, T::Out> where
R: Detokenize,
A: Into<Option<Address>>,
P: Tokenize,

Call constant function