Module web3::types

Web3 Types

Structs

Block
The block type returned from RPC calls. This is generic over a TX type.

Bytes
Raw bytes wrapper

CallRequest
Call contract request (eth_call / eth_estimateGas)

Filter
Filter

FilterBuilder
Filter Builder

H64
Uint serialization.

H128
Uint serialization.

H160
Uint serialization.

H256
Uint serialization.

H512
Uint serialization.

H520
Uint serialization.

H2048
Uint serialization.

Log
A log produced by a transaction.

Transaction
Description of a Transaction, pending or in the chain.

TransactionReceipt
"Receipt" of an executed transaction: details of its execution.

TransactionRequest
Send Transaction Parameters

U64
Uint serialization.

U256
Uint serialization.

Work
Miner's work package

Enums

BlockId
Block Identifier

BlockNumber
Block Number

TransactionCondition
Represents condition on minimum block number or block timestamp.

TransactionId
Transaction Identifier

Type Definitions

Address
Address

Index
Index in block