Module web3::types

Web3 Types

Structs

Block
The block type returned from RPC calls. This is generic over a TX type.

BlockHeader
The block header type returned from RPC calls.

BlockTrace
Ad-Hoc trace API type

Bytes
Raw bytes wrapper

CallRequest
Call contract request (eth_call / eth_estimateGas)

Filter
Filter

FilterBuilder
Filter Builder

H64
Unformatted binary data of fixed length.

H128
Unformatted binary data of fixed length.

H160
Unformatted binary data of fixed length.

H256
Unformatted binary data of fixed length.

H512
Unformatted binary data of fixed length.

H520
Unformatted binary data of fixed length.

H2048
Unformatted binary data of fixed length.

Log
A log produced by a transaction.

RawTransaction
Raw bytes of a signed, but not yet sent transaction

SyncInfo
Information about current blockchain syncing operations.

Trace
Trace-Filtering API trace type

TraceFilter
Trace filter

TraceFilterBuilder
Trace Filter Builder

Transaction
Description of a Transaction, pending or in the chain.

TransactionReceipt
"Receipt" of an executed transaction: details of its execution.

TransactionRequest
Send Transaction Parameters

TransactionTrace
Trace

U64
Little-endian large integer type

U128
Little-endian large integer type

U256
Little-endian large integer type

Work
Miner's work package

Enums

Action
Action

BlockId
Block Identifier

BlockNumber
Block Number

CallType
Call type.

Res
Response

SyncState
The current state of blockchain syncing operations.

TraceType
Description of the type of trace to make

TransactionCondition
Represents condition on minimum block number or block timestamp.

TransactionId
Transaction Identifier

Type Definitions

Address
Address

Index
Index in block